+++ avioserver/hardware_modules/motors.py
@@ -4,30 +4,33 @@ import Adafruit_PCA9685
 
 pwm = Adafruit_PCA9685.PCA9685()
 
-aerlon_channel = 0;
-aerlon_opposite_channel = 1;
+aerlon_channel = 0
+aerlon_opposite_channel = 1
 
 # Configure min and max servo pulse lengths
 servo_min = 150  # Min pulse length out of 4096
 servo_max = 600  # Max pulse length out of 4096
 
-aerlon_max = 3000;
+aerlon_max = 3000
+
 
 # Helper function to make setting a servo pulse width simpler.
 def set_servo_pulse(channel, pulse):
-    pulse_length = 1000000    # 1,000,000 us per second
-    pulse_length //= 60       # 60 Hz
+    pulse_length = 1000000  # 1,000,000 us per second
+    pulse_length //= 60  # 60 Hz
     print('{0}us per period'.format(pulse_length))
-    pulse_length //= 4096     # 12 bits of resolution
+    pulse_length //= 4096  # 12 bits of resolution
     print('{0}us per bit'.format(pulse_length))
     pulse *= 1000
     pulse //= pulse_length
     pwm.set_pwm(channel, 0, pulse)
 
+
 def set_aerlon(position):
-	pwm.set_pwm(aerlon_channel,0,position)
-	pwm.set_pwm(aerlon_opposite_channel,0,aerlon_max-position)
-	
+    pwm.set_pwm(aerlon_channel, 0, position)
+    pwm.set_pwm(aerlon_opposite_channel, 0, aerlon_max - position)
+
+
 # Set frequency to 60hz, good for servos.
 pwm.set_pwm_freq(60)
 
@@ -35,7 +38,7 @@ print('Moving servo on channel 0, press Ctrl-C to quit...')
 while True:
     # Move servo on channel O between extremes.
     pwm.set_pwm(1, 0, servo_min)
-    pwm.set_pwm(0, 0, servo_min)    
+    pwm.set_pwm(0, 0, servo_min)
     time.sleep(1)
     pwm.set_pwm(1, 0, servo_max)
     pwm.set_pwm(0, 0, servo_max)