+++ avioserver/views/controllers.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 from flask import Blueprint, request
 from avioserver.utils import ok_response, error_response
+from avioserver.hardware_modules.motor_communicator import MotorController
 
 avio_controllers = Blueprint("controllers", __name__)
+AvioMotorController = MotorController()
 
 
 @avio_controllers.route('/', methods=["GET"])
@@ -25,6 +27,7 @@ def error_index():
 
 @avio_controllers.route('/motors/wings', methods=["POST"])
 def control_wings():
+    AvioMotorController.set_aerlon(int(request.json["position"]))
     return ok_response(
         {
             "message": "Hello",